fix: add missing type to MenuItem
